Queue lesson

+++ app/Http/Controllers/JobController.php
@@ -2,15 +2,17 @@
 
 namespace App\Http\Controllers;
 
-use Illuminate\Http\Request;
+use App\Mail\JobPosted;
 use App\Models\Job;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Support\Facades\Mail;
 
 class JobController extends Controller
 {
     public function index()
     {
         $jobs = Job::with('employer')->latest()->simplePaginate(3);
- 
+
         return view('jobs.index', [
             'jobs' => $jobs
         ]);
@@ -32,13 +34,17 @@ class JobController extends Controller
             'title' => ['required', 'min:3'],
             'salary' => ['required']
         ]);
-    
-        Job::create([
+
+        $job = Job::create([
             'title' => request('title'),
             'salary' => request('salary'),
             'employer_id' => 1
         ]);
-    
+
+        Mail::to($job->employer->user)->queue(
+            new JobPosted($job)
+        );
+
         return redirect('/jobs');
     }
 
@@ -49,27 +55,27 @@ class JobController extends Controller
 
     public function update(Job $job)
     {
-        // authorize (On hold...)
-    request()->validate([
-        'title' => ['required', 'min:3'],
-        'salary' => ['required']
-    ]);
+        Gate::authorize('edit-job', $job);
 
-    $job->update([
-        'title' => request('title'),
-        'salary' => request('salary'),
-    ]);
+        request()->validate([
+            'title' => ['required', 'min:3'],
+            'salary' => ['required']
+        ]);
 
-    return redirect('/jobs/' . $job->id);
+        $job->update([
+            'title' => request('title'),
+            'salary' => request('salary'),
+        ]);
+
+        return redirect('/jobs/' . $job->id);
     }
 
     public function destroy(Job $job)
     {
-         // authorize (On hold...)
-    $job->delete();
+        Gate::authorize('edit-job', $job);
 
-    return redirect('/jobs');
+        $job->delete();
+
+        return redirect('/jobs');
     }
-
-
 }
